+++ mail/mail-common/src/draft.rs
@@ -54,7 +54,7 @@ pub mod recipients;
 pub(crate) mod send;
 
 use crate::draft::compose::{
-    DraftAddressChangeOutput, DraftAddressChangeRequest, encrypt_draft_body, get_signature,
+    DraftAddressChangeOutput, DraftAddressChangeRequest, encrypt_draft_body, get_full_signature,
     inject_dark_mode, patch_draft_with_reply_mode, prepare_html_reply, prepare_plain_text_reply,
 };
 pub use send::ScheduleSendOptions;
@@ -507,7 +507,7 @@ impl Draft {
             (None, DraftSyncStatus::Synced)
         } else if let Some(remote_id) = message.remote_id.clone() {
             debug!("Draft metadata has no pending changes, syncing.");
-            match Message::force_sync_message_and_body(context, remote_id, true).await {
+            match Message::force_sync_message_and_body(context, remote_id, true, tether).await {
                 Ok((message_new, decrypted)) => {
                     message = message_new;
 
@@ -648,7 +648,8 @@ impl Draft {
         address: &Address,
         mail_settings: &MailSettings,
     ) -> Self {
-        let body = compose::get_signature(address, mail_settings, mail_settings.draft_mime_type);
+        let body =
+            compose::get_full_signature(address, mail_settings, mail_settings.draft_mime_type);
         Self {
             metadata_id,
             sender: address.email.clone(),
@@ -845,7 +846,7 @@ impl Draft {
             MimeType::TextPlain
         };
 
-        let mut body = get_signature(address, mail_settings, mime_type);
+        let mut body = get_full_signature(address, mail_settings, mime_type);
 
         // If the message we are replying to is HTML we should also generate an HTML body for
         // replying even if the user has selected plain text as the default editing mode.
@@ -1003,7 +1004,7 @@ impl Draft {
     /// # Errors
     ///
     /// Returns error if the action failed to execute.
-    #[tracing::instrument(level=tracing::Level::DEBUG, skip(self,queue))]
+    #[tracing::instrument(level=tracing::Level::DEBUG, skip_all)]
     pub async fn save(
         &mut self,
         queue: &Queue,
@@ -1018,7 +1019,7 @@ impl Draft {
     /// # Errors
     ///
     /// Returns error if the action failed to execute.
-    #[tracing::instrument(level=tracing::Level::DEBUG, skip(self,queue))]
+    #[tracing::instrument(level=tracing::Level::DEBUG, skip_all)]
     pub async fn send(
         &mut self,
         queue: &Queue,
@@ -1035,7 +1036,7 @@ impl Draft {
     /// # Errors
     ///
     /// Returns error if the action failed to execute.
-    #[tracing::instrument(level=tracing::Level::DEBUG, skip(self,queue))]
+    #[tracing::instrument(level=tracing::Level::DEBUG, skip_all)]
     pub async fn schedule_send(
         &mut self,
         delivery_time: DateTime<Local>,
@@ -1052,7 +1053,7 @@ impl Draft {
     /// # Errors
     ///
     /// Returns error if the action failed to execute.
-    #[tracing::instrument(level=tracing::Level::DEBUG, skip(self,queue))]
+    #[tracing::instrument(level=tracing::Level::DEBUG, skip_all)]
     pub async fn discard(
         &self,
         queue: &Queue,
@@ -1900,7 +1901,7 @@ impl DraftAttachmentRemovalQueuer {
                     e => return Err(e.into()),
                 }
             }
-            metadata = metadata.with_dependency(action_id);
+            metadata = metadata.with_sequential_dependency(action_id);
         };
 
         Ok(queue