+++ mail/mail-common/src/user_context/events/messages.rs
@@ -1,17 +1,18 @@
-use crate::AppError;
+use crate::datatypes::MessageFlags;
 use crate::events::MessageEvent;
-use crate::models::Message;
+use crate::models::{DraftMetadata, Message};
+use crate::{AppError, user_context::events::subscriber::PostEventSyncData};
 use proton_core_common::events::Action;
-use proton_mail_ids::LocalMessageId;
+use proton_core_common::models::ModelIdExtension;
 use stash::params;
 use stash::stash::Bond;
-use tracing::{debug, warn};
+use tracing::warn;
 
 pub async fn handle_message_events(
     tx: &Bond<'_>,
     message_events: &[MessageEvent],
-) -> Result<Vec<LocalMessageId>, AppError> {
-    let mut ids = Vec::with_capacity(message_events.len());
+    data: &mut PostEventSyncData,
+) -> Result<(), AppError> {
     for message_event in message_events {
         message_event.action.log_entry(&message_event.remote_id);
         match message_event.action {
@@ -23,38 +24,73 @@ pub async fn handle_message_events(
                 .await?;
             }
             Action::Create => {
-                if let Some(message) = &message_event.message {
-                    let created =
-                        Message::create_or_update_messages_from_metadata(vec![message.clone()], tx)
-                            .await?;
-                    ids.extend(created);
-                } else {
-                    warn!("Received create message without message");
-                }
+                let Some(message) = &message_event.message else {
+                    warn!("Received update message without label");
+                    continue;
+                };
+
+                let created =
+                    Message::create_or_update_messages_from_metadata(vec![message.clone()], tx)
+                        .await?;
+                data.msg_for_prefetch.extend(created);
             }
             Action::Update | Action::UpdateFlags => {
-                if let Some(message) = &message_event.message {
-                    if !Message::can_update_from_event_loop(
-                        message_event.remote_id.clone(),
-                        message.flags.into(),
-                        tx,
-                    )
+                let Some(message) = &message_event.message else {
+                    warn!("Received update message without label");
+                    continue;
+                };
+
+                // Here the following cases can happen:
+                // 1. It's a draft, we don't have it open: Treat it as a normal message, update.
+                // 2. It's a draft, we have it open: Skip body and metadata updates because we
+                // don't have conflict resolution strategies in place
+                // 3. It _was_ a draft, we have it open, now it has been sent: We might have
+                // missed updates, let's do a full update.
+
+                if DraftMetadata::find_by_message_with_remote_id(message.id.clone(), tx)
                     .await?
-                    {
-                        debug!(
-                            "Skipping message update for {} due to draft metadata",
-                            message_event.remote_id
+                    .is_some()
+                {
+                    // We have a message that has been opened as a draft, but it is possible that
+                    // another session has sent this draft. Deleting the metadata at this point in
+                    // time can trigger the composer to display a collection of metadata not found errors
+                    // that can be very confusing for the user.
+                    // We let the update progress and the next action that executes for that
+                    // draft will trigger a failure and clean itself up.
+                    // It's possible that some messages will never properly clean up this way, but
+                    // this should happen very often and the associated metadata is not very large
+                    // with each draft. Correctly solving this requires knowledge of active composer
+                    // states on the rust side.
+
+                    let flags = MessageFlags::from(message.flags);
+                    if !(flags.is_schedule_send() || flags.is_sent()) {
+                        // Case 2.
+                        tracing::info!(
+                            "Skipping message update for {} because it's opened locally",
+                            message.id
                         );
                         continue;
                     }
-                    Message::create_or_update_messages_from_metadata(vec![message.clone()], tx)
-                        .await?;
-                } else {
-                    warn!("Received update message without label");
+
+                    tracing::info!(
+                        "Message {} has draft metadata but was already sent, update will be allowed",
+                        message.id
+                    );
+
+                    if let Some(local_id) =
+                        Message::remote_id_counterpart(message.id.clone(), tx).await?
+                    {
+                        // Case 3.
+                        // We delete the local message body so that it gets re-requested
+                        // whenever it gets open again. This is because we're skipping updates.
+                        Message::delete_message_body(local_id, tx).await?;
+                    }
                 }
+
+                Message::create_or_update_messages_from_metadata(vec![message.clone()], tx).await?;
             }
         }
     }
 
-    Ok(ids)
+    Ok(())
 }